eval-worker/pyproject.toml
[project]
name = "eval-worker"
version = "0.1.0"
description = "Background LLM-as-judge eval worker for AT."
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29",
    "openai>=1.30",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]
